feat #1210: move app-header into reuseable folder
