+++ packages/marketplace/src/components/ui/standalone-app-detail/__tests__/app-header.test.tsx
@@ -1,0 +1,17 @@
+import * as React from 'react'
+import { shallow } from 'enzyme'
+import { appDetailDataStub } from '@/sagas/__stubs__/app-detail'
+import AppHeader, { AppHeaderProps } from '../app-header'
+
+const mockProps: AppHeaderProps = {
+  appDetailData: {
+    ...appDetailDataStub.data,
+    apiKey: '',
+  },
+}
+
+describe('AppHeader', () => {
+  it('should match a snapshot', () => {
+    expect(shallow(<AppHeader {...mockProps} />)).toMatchSnapshot()
+  })
+})
