+++ packages/marketplace/src/styles/blocks/standalone-app-detail.scss
@@ -1,0 +1,43 @@
+@import '../base/colors.scss';
+
+.appIconContainer {
+  margin-left: 0;
+  margin-right: 2rem;
+  width: 128px;
+  height: 128px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background-color: $grey-lighter;
+  border-radius: 50%;
+  margin: 0 auto;
+}
+
+.verifyByReapitContainer {
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+  margin-top: 0.5rem;
+  margin-bottom: 1rem;
+
+  .check {
+    /*should be px regard their base px (rem) else looks really ugly*/
+    margin-right: 5px;
+  }
+}
+
+.appHeaderFeaturedImage {
+  max-width: 40%;
+  margin-left: 1rem;
+}
+
+.appHeader {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+
+  h2.appName {
+    font-family: 'Museo Sans W03_300', 'Roboto', Helvetica, Arial, sans-serif;
+    margin-bottom: 0;
+  }
+}
