+++ packages/marketplace/src/styles/pages/client-app-detail.scss
@@ -1,34 +1,5 @@
 @import '../base/colors.scss';
 
-.appHeader {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-
-  h2.appName {
-    font-family: 'Museo Sans W03_300', 'Roboto', Helvetica, Arial, sans-serif;
-    margin-bottom: 0;
-  }
-}
-
-.verifyByReapitContainer {
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  margin-top: 0.5rem;
-  margin-bottom: 1rem;
-
-  .check {
-    /*should be px regard their base px (rem) else looks really ugly*/
-    margin-right: 5px;
-  }
-}
-
-.appIconContainer {
-  margin-left: 0;
-  margin-right: 2rem;
-}
-
 .check {
   color: #a0c862;
 }
@@ -62,8 +33,4 @@
   list-style-position: inside;
 }
 
-.appHeaderFeaturedImage {
-  max-width: 40%;
-  margin-left: 1rem;
-}
