+++ packages/marketplace/src/styles/pages/developer-app-detail.scss
@@ -50,16 +50,6 @@
   padding: 15px;
 }
 
-.appIconContainer {
-  width: 128px;
-  height: 128px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background-color: $grey-lighter;
-  border-radius: 50%;
-  margin: 0 auto;
-}
 
 .renderAsideStatusContainer {
   display: flex;
